backend/requirements.txt
django
psycopg[binary]
python-dotenv
djangorestframework
markdown
PyYAML
uritemplate
pygments
django-filter
pandas